v2

+++ 2019-09-11/precision.cpp
@@ -3,15 +3,15 @@
 
 int main(void)
 {
-  float eps=1.0;
-  float one=1.0;
+  double eps=1.0;
+  double one=1.0;
 
   for(int i=1;i<300;i++)
     {
-      eps=eps/2;
+      eps=eps/2.0;
       
       one=1.0+eps;
-       printf ("%5d %15.10e %15.10e\n", i, one,eps);
+       printf ("%5d %25.20e %25.20e\n", i, one,eps);
   
     }
